backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.10.3
pydantic-settings>=2.7.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
